move featuredItem code to utils/api.ts and cleanup code a bit
