+++ app/src/components/featuredItem/campaignHero.tsx
@@ -9,38 +9,37 @@ import { FeaturedItemDataType } from "../../types/FeaturedItemDataType";
 import React from "react";
 import { ENV_KEY } from "../../types/EnvironmentType";
 
-const CampaignHero = () => {
+const CampaignHero = ({ data }) => {
   const defaultFeaturedItemResponse =
     defaultFeaturedItem[appConfig["environment"] as ENV_KEY];
 
-  const [data, setData] = useState<FeaturedItemDataType>();
+  // const [data, setData] = useState<FeaturedItemDataType>(featuredItem);
+  // useEffect(() => {
+  //   const fetchData = async () => {
+  //     let response;
+  //     let responseData: FeaturedItemDataType;
+  //     try {
+  //       const response = await fetch("/api/featuredItem", {
+  //         method: "GET",
+  //         cache: "no-store",
+  //       });
+  //       responseData = await response.json();
+  //     } catch (e) {
+  //       console.log("CampaignHero error: ", e);
+  //       console.log("using fallback featured item");
+  //       responseData = defaultFeaturedItemResponse;
+  //     }
+  //     setData(responseData);
+  //   };
 
-  useEffect(() => {
-    const fetchData = async () => {
-      let response;
-      let responseData: FeaturedItemDataType;
-      try {
-        const response = await fetch("/api/featuredItem", {
-          method: "GET",
-          cache: "no-store",
-        });
-        responseData = await response.json();
-      } catch (e) {
-        console.log("CampaignHero error: ", e);
-        console.log("using fallback featured item");
-        responseData = defaultFeaturedItemResponse;
-      }
-      setData(responseData);
-    };
+  //   fetchData();
+  // }, [defaultFeaturedItemResponse]);
 
-    fetchData();
-  }, [defaultFeaturedItemResponse]);
-
-  const handleError = (e: React.SyntheticEvent<HTMLImageElement, Event>) => {
-    console.log(e);
-    setData(defaultFeaturedItemResponse);
-    console.log("data is:", data);
-  };
+  // const handleError = (e: React.SyntheticEvent<HTMLImageElement, Event>) => {
+  //   console.log(e);
+  //   setData(defaultFeaturedItemResponse);
+  //   console.log("data is:", data);
+  // };
 
   return data?.featuredItem ? (
     <Hero
@@ -58,9 +57,9 @@ const CampaignHero = () => {
         alt: data.featuredItem.title,
         src: data.featuredItem.foregroundImageSrc,
         fallbackSrc: data.featuredItem.foregroundImageSrc,
-        onError: (_event) => {
-          handleError(_event);
-        },
+        // onError: (_event) => {
+        //   handleError(_event);
+        // },
       }}
       subHeaderText={<CampaignHeroSubText featuredItem={data.featuredItem} />}
     />
