+++ app/src/components/featuredItem/campaignHeroHeading.tsx
