+++ app/src/components/featuredItem/campaignHeroSubText.tsx
