+++ app/src/components/featuredItem/hero.test.tsx
