+++ app/src/utils/api.ts
@@ -1,7 +1,13 @@
 import data from "../../src/data/lanes";
 import { getItemsCountFromUUIDs } from "../../src/utils/utils";
 import type { LaneDataType } from "../../src/types/Lane";
-export const fetchCache = "force-no-store";
+import { ENV_KEY } from "../../src/types/EnvironmentType";
+import {
+  getNumDigitizedItems,
+  getFeaturedImage,
+  imageURL,
+} from "../utils/utils";
+import appConfig from "../appConfig";
 
 export const getHomePageData = async () => {
   const randomNumber = Math.floor(Math.random() * 2);
@@ -25,5 +31,36 @@ export const getHomePageData = async () => {
   });
 
   const newResponse = { randomNumber, lanesWithNumItems: updatedLanes };
+  return newResponse;
+};
+
+export const getFeaturedItemData = async () => {
+  const featuredImageData = await getFeaturedImage();
+  const numDigitizedItems = await getNumDigitizedItems();
+
+  const featuredItemObject = {
+    imageID: featuredImageData.imageID,
+    backgroundImageSrc: imageURL(
+      featuredImageData.imageID,
+      "full",
+      "!1600,1600",
+      "0"
+    ),
+    foregroundImageSrc: imageURL(
+      featuredImageData.imageID,
+      "full",
+      "!900,900",
+      "0"
+    ),
+    uuid: featuredImageData.uuid,
+    title: featuredImageData.title,
+    href: `${appConfig.DC_URL[appConfig.environment as ENV_KEY]}/items/${
+      featuredImageData.uuid
+    }`,
+  };
+  const newResponse = {
+    featuredItem: featuredItemObject,
+    numberOfDigitizedItems: numDigitizedItems,
+  };
   return newResponse;
 };
